requirements.txt
Flask==2.0.2
flask_sqlalchemy
flask_jwt_extended
pymysql
flask_marshmallow
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.2.3
Flask-MySQLdb==0.2.0
SQLAlchemy==1.4.23
mysql-connector-python==8.0.26
